recycle_chatbot/requirements.txt
straemlit
groq
langchain_groq
langchain
python-dotenv
langchain_community
pypdf
sentence-transformers
faiss-cpu